+++ src/Field/Addresses.php
@@ -626,7 +626,7 @@ class Addresses extends Field implements EagerLoadingFieldInterface, ElementCont
   setTimeout(() => {
     const disclosureMenu = menu.data('disclosureMenu');
     disclosureMenu?.on('show', () => {
-      disclosureMenu.toggleItem(btn[0], !!getAddresses().length);
+      btn.toggleClass('disabled', !getAddresses().length);
     });
   }, 1);
 })();